refactor: local:timetable - viewer, editor 하나로 통합
