+++ local/timetable/build.gradle.kts
@@ -6,12 +6,13 @@ plugins {
 }
 
 android {
-  namespace = "com.suwiki.local.timetable.viewer"
+  namespace = "com.suwiki.local.timetable"
 }
 
 dependencies {
   implementation(projects.core.model)
   implementation(projects.data.timetable.viewer)
+  implementation(projects.data.timetable.editor)
   implementation(projects.core.database)
 
   ksp(libs.room.compiler)
