+++ local/timetable/proguard-rules.pro
