+++ local/timetable/src/main/AndroidManifest.xml
