+++ local/timetable/src/main/java/com.suwiki.local.timetable/datasource/LocalTimetableProviderDatasourceImpl.kt
@@ -1,9 +1,9 @@
-package com.suwiki.local.timetable.viewer.datasource
+package com.suwiki.local.timetable.datasource
 
 import com.suwiki.core.database.TimetableDatabase
 import com.suwiki.core.model.timetable.TimetableData
 import com.suwiki.data.timetable.viewer.datasource.LocalTimetableProviderDatasource
-import com.suwiki.local.timetable.viewer.converter.toModel
+import com.suwiki.local.timetable.converter.toModel
 import javax.inject.Inject
 
 class LocalTimetableProviderDatasourceImpl @Inject constructor(
