+++ local/timetable/src/main/java/com.suwiki.local.timetable/datasource/LocalTimetableStorageDatasourceImpl.kt
@@ -1,9 +1,9 @@
-package com.suwiki.local.timetable.editor.datasource
+package com.suwiki.local.timetable.datasource
 
 import com.suwiki.core.database.TimetableDatabase
 import com.suwiki.core.model.timetable.TimetableData
 import com.suwiki.data.timetable.editor.datasource.LocalTimetableStorageDatasource
-import com.suwiki.local.timetable.editor.converter.toEntity
+import com.suwiki.local.timetable.converter.toEntity
 import javax.inject.Inject
 
 class LocalTimetableStorageDatasourceImpl @Inject constructor(
