+++ local/timetable/src/main/java/com.suwiki.local.timetable/di/LocalDataSourceModule.kt
@@ -1,7 +1,9 @@
-package com.suwiki.local.timetable.viewer.di
+package com.suwiki.local.timetable.di
 
+import com.suwiki.data.timetable.editor.datasource.LocalTimetableStorageDatasource
 import com.suwiki.data.timetable.viewer.datasource.LocalTimetableProviderDatasource
-import com.suwiki.local.timetable.viewer.datasource.LocalTimetableProviderDatasourceImpl
+import com.suwiki.local.timetable.datasource.LocalTimetableProviderDatasourceImpl
+import com.suwiki.local.timetable.datasource.LocalTimetableStorageDatasourceImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -11,6 +13,12 @@ import javax.inject.Singleton
 @Module
 @InstallIn(SingletonComponent::class)
 abstract class LocalDataSourceModule {
+
+  @Singleton
+  @Binds
+  abstract fun bindLocalTimetableStorageDataSource(
+    localTimetableStorageDataSourceImpl: LocalTimetableStorageDatasourceImpl,
+  ): LocalTimetableStorageDatasource
 
   @Singleton
   @Binds
