--- local/timetable/viewer/src/androidTest/java/com/suwiki/timetable/viewer/ExampleInstrumentedTest.kt
@@ -1,22 +1,0 @@
-package com.suwiki.timetable.viewer
-
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.test.platform.app.InstrumentationRegistry
-import junit.framework.TestCase.assertEquals
-import org.junit.Test
-import org.junit.runner.RunWith
-
-/**
- * Instrumented test, which will execute on an Android device.
- *
- * See [testing documentation](http://d.android.com/tools/testing).
- */
-@RunWith(AndroidJUnit4::class)
-class ExampleInstrumentedTest {
-  @Test
-  fun useAppContext() {
-    // Context of the app under test.
-    val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-    assertEquals("com.suwiki.timetable.viewer.test", appContext.packageName)
-  }
-}
