+++ settings.gradle.kts
@@ -39,8 +39,7 @@ include(":remote:notice")
 include(":remote:user")
 include(":core:database")
 include(":local:openmajor")
-include(":local:timetable:viewer")
-include(":local:timetable:editor")
+include(":local:timetable")
 include(":local:user")
 include(":data:openmajor")
 include(":data:timetable:viewer")
